% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRD_utility.R
\name{utilsCalculateDistance}
\alias{utilsCalculateDistance}
\title{utilsCalculateDistance}
\usage{
utilsCalculateDistance(df, nameCol, refCol)
}
\arguments{
\item{df}{A DataFrame.}

\item{nameCol}{The current Column of the iteration.}

\item{refCol}{The reference Column of the dataFrame.}
}
\value{
Returns a new \code{df} that has a Distance Column based on the \code{nameCol}.
}
\description{
Calculates the Manhattan-distance between two rankings and inserts it into the DataFrame after the first column.
}
\examples{
SRDInput <- data.frame(
A=c(32, 52, 44, 44, 47),
B=c(73, 75, 65, 76, 70),
C=c(60, 59, 57, 55, 60),
D=c(35, 24, 44, 83, 47),
E=c(41, 52, 46, 50, 65))
nameCol <- "A"
refCol <- "B"
 for(i in names(SRDInput)){ SRDInput <- rSRD:::utilsCalculateRank(SRDInput,i)}
rSRD::utilsCalculateDistance(SRDInput,nameCol,refCol)
}
\author{
Ali Tugay Sen
}
